refactor: add OpenAI optional egg
