+++ haystack/nodes/answer_generator/openai.py
@@ -5,8 +5,14 @@ import json
 import logging
 import requests
 
-import tiktoken
-from tiktoken import Encoding
+
+try:
+    import tiktoken
+    from tiktoken import Encoding
+except (ImportError, ModuleNotFoundError) as ie:
+    from haystack.utils.import_utils import _optional_component_not_installed
+
+    _optional_component_not_installed(__name__, "openai", ie)
 
 from haystack.nodes.answer_generator import BaseGenerator
 from haystack import Document
