+++ haystack/nodes/retriever/__init__.py
@@ -1,17 +1,17 @@
 from haystack.nodes.retriever.base import BaseRetriever
 from haystack.nodes.retriever.dense import (
-    DenseRetriever,
     DensePassageRetriever,
+    DenseRetriever,
     EmbeddingRetriever,
     MultihopEmbeddingRetriever,
     TableTextRetriever,
 )
+from haystack.nodes.retriever.multimodal import MultiModalRetriever
 from haystack.nodes.retriever.sparse import (
     BM25Retriever,
-    ElasticsearchRetriever,
     ElasticsearchFilterOnlyRetriever,
+    ElasticsearchRetriever,
     FilterRetriever,
     TfidfRetriever,
 )
 from haystack.nodes.retriever.text2sparql import Text2SparqlRetriever
-from haystack.nodes.retriever.multimodal import MultiModalRetriever
