+++ haystack/nodes/retriever/_embedding_encoder.py
@@ -1,13 +1,10 @@
 import json
 import logging
-from abc import abstractmethod
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union, Type
 
 import numpy as np
 import requests
-import tiktoken
-from tiktoken import Encoding
 import torch
 from sentence_transformers import InputExample
 from torch.utils.data import DataLoader
@@ -15,110 +12,23 @@ from torch.utils.data.sampler import SequentialSampler
 from tqdm.auto import tqdm
 from transformers import AutoModel, AutoTokenizer
 
-from haystack.document_stores.base import BaseDocumentStore
-from haystack.errors import CohereError, OpenAIError, OpenAIRateLimitError
+from haystack.errors import CohereError
 from haystack.modeling.data_handler.dataloader import NamedDataLoader
 from haystack.modeling.data_handler.dataset import convert_features_to_dataset, flatten_rename
 from haystack.modeling.infer import Inferencer
+
 from haystack.nodes.retriever._losses import _TRAINING_LOSSES
 from haystack.schema import Document
+from haystack.utils.import_utils import safe_import, MissingDependency
 from haystack.utils.reflection import retry_with_exponential_backoff
+from ._base_encoder import _BaseEmbeddingEncoder
 
 if TYPE_CHECKING:
     from haystack.nodes.retriever import EmbeddingRetriever
 
+_OpenAIEmbeddingEncoder = safe_import("haystack.nodes.retriever._openai_encoder", "_OpenAIEmbeddingEncoder", "openai")
+
 logger = logging.getLogger(__name__)
-
-
-class _BaseEmbeddingEncoder:
-    @abstractmethod
-    def embed_queries(self, queries: List[str]) -> np.ndarray:
-        """
-        Create embeddings for a list of queries.
-
-        :param queries: List of queries to embed.
-        :return: Embeddings, one per input query, shape: (queries, embedding_dim)
-        """
-        pass
-
-    @abstractmethod
-    def embed_documents(self, docs: List[Document]) -> np.ndarray:
-        """
-        Create embeddings for a list of documents.
-
-        :param docs: List of documents to embed.
-        :return: Embeddings, one per input document, shape: (documents, embedding_dim)
-        """
-        pass
-
-    def train(
-        self,
-        training_data: List[Dict[str, Any]],
-        learning_rate: float = 2e-5,
-        n_epochs: int = 1,
-        num_warmup_steps: Optional[int] = None,
-        batch_size: int = 16,
-    ):
-        """
-        Trains or adapts the underlying embedding model.
-
-        Each training data example is a dictionary with the following keys:
-
-        * question: The question string.
-        * pos_doc: Positive document string (the document containing the answer).
-        * neg_doc: Negative document string (the document that doesn't contain the answer).
-        * score: The score margin the answer must fall within.
-
-
-        :param training_data: The training data in a dictionary format. Required.
-        :type training_data: List[Dict[str, Any]]
-        :param learning_rate: The speed at which the model learns. Required. We recommend that you leave the default `2e-5` value.
-        :type learning_rate: float
-        :param n_epochs: The number of epochs (complete passes of the training data through the algorithm) that you want the model to go through. Required.
-        :type n_epochs: int
-        :param num_warmup_steps: The number of warmup steps for the model. Warmup steps are epochs when the learning rate is very low. You can use them at the beginning of the training to prevent early overfitting of your model. Required.
-        :type num_warmup_steps: int
-        :param batch_size: The batch size to use for the training. Optional. The default values is 16.
-        :type batch_size: int (optional)
-        """
-        pass
-
-    def save(self, save_dir: Union[Path, str]):
-        """
-        Save the model to the directory you specify.
-
-        :param save_dir: The directory where the model is saved. Required.
-        :type save_dir: Union[Path, str]
-        """
-        pass
-
-    def _check_docstore_similarity_function(self, document_store: BaseDocumentStore, model_name: str):
-        """
-        Check that document_store uses a similarity function
-        compatible with the embedding model
-        """
-        if "sentence-transformers" in model_name.lower():
-            model_similarity = None
-            if "-cos-" in model_name.lower():
-                model_similarity = "cosine"
-            elif "-dot-" in model_name.lower():
-                model_similarity = "dot_product"
-
-            if model_similarity is not None and document_store.similarity != model_similarity:
-                logger.warning(
-                    "You seem to be using %s model with the %s function instead of the recommended %s. "
-                    "This can be set when initializing the DocumentStore",
-                    model_name,
-                    document_store.similarity,
-                    model_similarity,
-                )
-        elif "dpr" in model_name.lower() and document_store.similarity != "dot_product":
-            logger.warning(
-                "You seem to be using a DPR model with the %s function. "
-                "We recommend using dot_product instead. "
-                "This can be set when initializing the DocumentStore",
-                document_store.similarity,
-            )
 
 
 class _DefaultEmbeddingEncoder(_BaseEmbeddingEncoder):
@@ -394,104 +304,6 @@ class _RetribertEmbeddingEncoder(_BaseEmbeddingEncoder):
         )
 
 
-class _OpenAIEmbeddingEncoder(_BaseEmbeddingEncoder):
-    def __init__(self, retriever: "EmbeddingRetriever"):
-        # See https://beta.openai.com/docs/guides/embeddings for more details
-        self.url = "https://api.openai.com/v1/embeddings"
-        self.api_key = retriever.api_key
-        self.batch_size = min(64, retriever.batch_size)
-        self.progress_bar = retriever.progress_bar
-        model_class: str = next(
-            (m for m in ["ada", "babbage", "davinci", "curie"] if m in retriever.embedding_model), "babbage"
-        )
-        self.tokenizer: Optional[Encoding] = None
-        self._setup_encoding_models(model_class, retriever.embedding_model, retriever.max_seq_len)
-
-    def _setup_encoding_models(self, model_class: str, model_name: str, max_seq_len: int):
-        """
-        Setup the encoding models for the retriever.
-        """
-        # new generation of embedding models (December 2022), we need to specify the full name
-        if "text-embedding" in model_name:
-            self.query_encoder_model = model_name
-            self.doc_encoder_model = model_name
-            self.max_seq_len = min(8191, max_seq_len)
-            tokenizer = "cl100k_base"
-        else:
-            self.query_encoder_model = f"text-search-{model_class}-query-001"
-            self.doc_encoder_model = f"text-search-{model_class}-doc-001"
-            self.max_seq_len = min(2046, max_seq_len)
-            tokenizer = "gpt2"
-
-        self.tokenizer = tiktoken.get_encoding(tokenizer)
-
-    def _ensure_text_limit(self, text: str) -> str:
-        """
-        Ensure that length of the text is within the maximum length of the model.
-        OpenAI embedding models have a limit of 2048 tokens
-        """
-        if self.tokenizer is None:
-            raise ValueError("Tokenizer is not initialized. Please call _setup_encoding_models first.")
-
-        tokenized_payload = self.tokenizer.encode(text)
-        return self.tokenizer.decode(tokenized_payload[: self.max_seq_len])
-
-    @retry_with_exponential_backoff(backoff_in_seconds=10, max_retries=5)
-    def embed(self, model: str, text: List[str]) -> np.ndarray:
-        payload = {"model": model, "input": text}
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-        response = requests.request("POST", self.url, headers=headers, data=json.dumps(payload), timeout=30)
-        res = json.loads(response.text)
-
-        if response.status_code != 200:
-            openai_error: OpenAIError
-            if response.status_code == 429:
-                openai_error = OpenAIRateLimitError(f"API rate limit exceeded: {response.text}")
-            else:
-                openai_error = OpenAIError(
-                    f"OpenAI returned an error.\n"
-                    f"Status code: {response.status_code}\n"
-                    f"Response body: {response.text}",
-                    status_code=response.status_code,
-                )
-            raise openai_error
-
-        unordered_embeddings = [(ans["index"], ans["embedding"]) for ans in res["data"]]
-        ordered_embeddings = sorted(unordered_embeddings, key=lambda x: x[0])
-        generated_embeddings = [emb[1] for emb in ordered_embeddings]
-        return np.array(generated_embeddings)
-
-    def embed_batch(self, model: str, text: List[str]) -> np.ndarray:
-        all_embeddings = []
-        for i in tqdm(
-            range(0, len(text), self.batch_size), disable=not self.progress_bar, desc="Calculating embeddings"
-        ):
-            batch = text[i : i + self.batch_size]
-            batch_limited = [self._ensure_text_limit(content) for content in batch]
-            generated_embeddings = self.embed(model, batch_limited)
-            all_embeddings.append(generated_embeddings)
-        return np.concatenate(all_embeddings)
-
-    def embed_queries(self, queries: List[str]) -> np.ndarray:
-        return self.embed_batch(self.query_encoder_model, queries)
-
-    def embed_documents(self, docs: List[Document]) -> np.ndarray:
-        return self.embed_batch(self.doc_encoder_model, [d.content for d in docs])
-
-    def train(
-        self,
-        training_data: List[Dict[str, Any]],
-        learning_rate: float = 2e-5,
-        n_epochs: int = 1,
-        num_warmup_steps: Optional[int] = None,
-        batch_size: int = 16,
-    ):
-        raise NotImplementedError(f"Training is not implemented for {self.__class__}")
-
-    def save(self, save_dir: Union[Path, str]):
-        raise NotImplementedError(f"Saving is not implemented for {self.__class__}")
-
-
 class _CohereEmbeddingEncoder(_BaseEmbeddingEncoder):
     def __init__(self, retriever: "EmbeddingRetriever"):
         # See https://docs.cohere.ai/embed-reference/ for more details
@@ -561,7 +373,7 @@ class _CohereEmbeddingEncoder(_BaseEmbeddingEncoder):
         raise NotImplementedError(f"Saving is not implemented for {self.__class__}")
 
 
-_EMBEDDING_ENCODERS: Dict[str, Callable] = {
+_EMBEDDING_ENCODERS: Dict[str, Union[Type[_BaseEmbeddingEncoder], Type[MissingDependency]]] = {
     "farm": _DefaultEmbeddingEncoder,
     "transformers": _DefaultEmbeddingEncoder,
     "sentence_transformers": _SentenceTransformersEmbeddingEncoder,
