+++ haystack/nodes/retriever/_openai_encoder.py
@@ -1,0 +1,123 @@
+import json
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
+
+import numpy as np
+import requests
+
+from tqdm.auto import tqdm
+
+try:
+    import tiktoken
+except (ImportError, ModuleNotFoundError) as ie:
+    from haystack.utils.import_utils import _optional_component_not_installed
+
+    _optional_component_not_installed(__name__, "openai", ie)
+
+
+from haystack.errors import OpenAIError, OpenAIRateLimitError
+from haystack.nodes.retriever._base_encoder import _BaseEmbeddingEncoder
+from haystack.schema import Document
+from haystack.utils.reflection import retry_with_exponential_backoff
+
+
+if TYPE_CHECKING:
+    from haystack.nodes.retriever import EmbeddingRetriever
+
+
+class _OpenAIEmbeddingEncoder(_BaseEmbeddingEncoder):
+    def __init__(self, retriever: "EmbeddingRetriever"):
+        # See https://beta.openai.com/docs/guides/embeddings for more details
+        self.url = "https://api.openai.com/v1/embeddings"
+        self.api_key = retriever.api_key
+        self.batch_size = min(64, retriever.batch_size)
+        self.progress_bar = retriever.progress_bar
+        model_class: str = next(
+            (m for m in ["ada", "babbage", "davinci", "curie"] if m in retriever.embedding_model), "babbage"
+        )
+        self.tokenizer: Optional[tiktoken.Encoding] = None
+        self._setup_encoding_models(model_class, retriever.embedding_model, retriever.max_seq_len)
+
+    def _setup_encoding_models(self, model_class: str, model_name: str, max_seq_len: int):
+        """
+        Setup the encoding models for the retriever.
+        """
+        # new generation of embedding models (December 2022), we need to specify the full name
+        if "text-embedding" in model_name:
+            self.query_encoder_model = model_name
+            self.doc_encoder_model = model_name
+            self.max_seq_len = min(8191, max_seq_len)
+            tokenizer = "cl100k_base"
+        else:
+            self.query_encoder_model = f"text-search-{model_class}-query-001"
+            self.doc_encoder_model = f"text-search-{model_class}-doc-001"
+            self.max_seq_len = min(2046, max_seq_len)
+            tokenizer = "gpt2"
+
+        self.tokenizer = tiktoken.get_encoding(tokenizer)
+
+    def _ensure_text_limit(self, text: str) -> str:
+        """
+        Ensure that length of the text is within the maximum length of the model.
+        OpenAI embedding models have a limit of 2048 tokens
+        """
+        if self.tokenizer is None:
+            raise ValueError("Tokenizer is not initialized. Please call _setup_encoding_models first.")
+
+        tokenized_payload = self.tokenizer.encode(text)
+        return self.tokenizer.decode(tokenized_payload[: self.max_seq_len])
+
+    @retry_with_exponential_backoff(backoff_in_seconds=10, max_retries=5)
+    def embed(self, model: str, text: List[str]) -> np.ndarray:
+        payload = {"model": model, "input": text}
+        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
+        response = requests.request("POST", self.url, headers=headers, data=json.dumps(payload), timeout=30)
+        res = json.loads(response.text)
+
+        if response.status_code != 200:
+            openai_error: OpenAIError
+            if response.status_code == 429:
+                openai_error = OpenAIRateLimitError(f"API rate limit exceeded: {response.text}")
+            else:
+                openai_error = OpenAIError(
+                    f"OpenAI returned an error.\n"
+                    f"Status code: {response.status_code}\n"
+                    f"Response body: {response.text}",
+                    status_code=response.status_code,
+                )
+            raise openai_error
+
+        unordered_embeddings = [(ans["index"], ans["embedding"]) for ans in res["data"]]
+        ordered_embeddings = sorted(unordered_embeddings, key=lambda x: x[0])
+        generated_embeddings = [emb[1] for emb in ordered_embeddings]
+        return np.array(generated_embeddings)
+
+    def embed_batch(self, model: str, text: List[str]) -> np.ndarray:
+        all_embeddings = []
+        for i in tqdm(
+            range(0, len(text), self.batch_size), disable=not self.progress_bar, desc="Calculating embeddings"
+        ):
+            batch = text[i : i + self.batch_size]
+            batch_limited = [self._ensure_text_limit(content) for content in batch]
+            generated_embeddings = self.embed(model, batch_limited)
+            all_embeddings.append(generated_embeddings)
+        return np.concatenate(all_embeddings)
+
+    def embed_queries(self, queries: List[str]) -> np.ndarray:
+        return self.embed_batch(self.query_encoder_model, queries)
+
+    def embed_documents(self, docs: List[Document]) -> np.ndarray:
+        return self.embed_batch(self.doc_encoder_model, [d.content for d in docs])
+
+    def train(
+        self,
+        training_data: List[Dict[str, Any]],
+        learning_rate: float = 2e-5,
+        n_epochs: int = 1,
+        num_warmup_steps: Optional[int] = None,
+        batch_size: int = 16,
+    ):
+        raise NotImplementedError(f"Training is not implemented for {self.__class__}")
+
+    def save(self, save_dir: Union[Path, str]):
+        raise NotImplementedError(f"Saving is not implemented for {self.__class__}")
