+++ pyproject.toml
@@ -210,7 +210,7 @@ dev = [
   "requests-cache",
 ]
 all = [
-  "farm-haystack[docstores,audio,crawler,preprocessing,ocr,ray,dev,onnx,beir]",
+  "farm-haystack[docstores,audio,crawler,preprocessing,ocr,ray,dev,onnx,beir,openai]",
 ]
 all-gpu = [
   # beir is incompatible with faiss-gpu: https://github.com/beir-cellar/beir/issues/71
